requirements.txt
awsebcli==3.2
bcdoc==0.13.0
boto==2.36.0
botocore==0.99.0
cement==2.4.0
Django==1.7.7
django-crispy-forms==1.4.0
django-parler==1.3
django-session-security==2.2.3
django-storages==1.1.8
django-tag-parser==2.0.1
docutils==0.12
future==0.14.3
jmespath==0.6.1
pyasn1==0.1.7
Pygments==2.0.2
python-dateutil==2.4.1
pytz==2015.2
PyYAML==3.11
six==1.9.0
